Another stab at EffectHandler
This time an effect handler takes a dispatch function instead of
returning an observable/promise. This is closer to what Mobius does and
makes for a better API.

The effects our now also fully testable as they can be anything
defined by the user. It should also be possible to define a effect
handler working on observables for full use of RxJS in the effect
handling.